fix somedatadef.php
